add testing for customer claim and change customer claim data type that double tp Big Decimal

+++ src/main/java/com/app/smartdrive/api/dto/customer/request/ClaimRequestDTO.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -18,7 +19,7 @@ public class ClaimRequestDTO {
 
     private Long creqEntityId;
 
-    private Double cuclEventPrice;
+    private BigDecimal cuclEventPrice;
 
-    private Double cuclSubtotal;
+    private BigDecimal cuclSubtotal;
 }
